Add ability to display LogBlock mob kill coordinates
This is a bit of a hack job, using slight modifications to the existing
block edit detecting functions and slipping in kill message appropriate
things where necessary.  This led to non-blocks in blocks.yml and so
forth, but it works rather nicely.  Models were kept simple and cuboid
but are quite distinguishable based on color and size.  I'd be happy to
make any changes based on recommendations from @totemo and please
consider this a, while perfectly working, rather rough-around-the-edges
first draft.